Working Imbalance Price tests.

+++ src/Shared/Messages/ElexonImbalancePriceItem.cs
@@ -4,6 +4,6 @@ public class ElexonImbalancePriceItem
 {
     public string SettlementDate { get; init; }
     public int SettlementPeriod { get; init; }
-    public double ImbalancePrice { get; init; }
+    public decimal ImbalancePrice { get; init; }
     public string StartDateTimeUTC { get; init; } 
 }
